+++ src/core/factory.ts
@@ -52,6 +52,7 @@ const DIALECT_LOADERS: Record<DialectType, () => Promise<{ createDialect: () => 
   sybase:      () => import(/* webpackIgnore: true */ /* @vite-ignore */ '../dialects/sybase.dialect.js'),
   duckdb:      () => import(/* webpackIgnore: true */ /* @vite-ignore */ '../dialects/duckdb.dialect.js'),
   firestore:   () => import(/* webpackIgnore: true */ /* @vite-ignore */ '../dialects/firestore.dialect.js'),
+  firebird:    () => import(/* webpackIgnore: true */ /* @vite-ignore */ '../dialects/firebird.dialect.js'),
 };
 
 /**